RecipeRunCycle now public, consolidate and limit access in WorkingDirectoryExecutionContext
